pyproject.toml
[project]
name = "veryessence"
version = "0.8.0"
description = "VeryEssence automates natural language analysis."
authors = [
    { name = "nos-tromo" }
]
requires-python = ">=3.11"
dependencies = [
    "accelerate>=1.1.1",
    "arabic-reshaper>=3.0.0",
    "bertopic>=0.16.4",
    "flask>=3.1.0",
    "huggingface-hub>=0.26.2",
    "kaleido>=0.4.1",
    "keybert>=0.8.5",
    "langchain>=0.3.7",
    "langchain-community>=0.3.7",
    "llama-cpp-python>=0.3.1",
    "matplotlib>=3.9.2",
    "networkx>=3.4.2",
    "nltk>=3.9.1",
    "numpy>=1.26.4",
    "openai-whisper>=20240930",
    "openpyxl>=3.1.5",
    "pandas>=2.2.3",
    "pillow>=11.0.0",
    "plotly>=5.24.1",
    "pyannote-audio>=3.3.2",
    "pycountry>=24.6.1",
    "pydub>=0.25.1",
    "python-bidi>=0.6.3",
    "python-dotenv>=1.0.1",
    "reportlab>=4.2.5",
    "sacremoses>=0.1.1",
    "scikit-learn>=1.5.2",
    "sentence-transformers>=3.3.0",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "transformers>=4.46.2",
    "umap-learn>=0.5.7",
    "wordcloud>=1.9.4",
]

[project.optional-dependencies]
cli = [
    "ruff>=0.7.4",
    "yt-dlp>=2024.11.4",
]

[project.urls]
Repository = "https://github.com/nos-tromo/VeryEssence.git"
